fix following and commenting bug when requiring sign in for signed in users

+++ app/assets/javascripts/utils/followable.js
@@ -15,31 +15,35 @@ MediumClone.Mixins.Followable = {
     var type
     var thisModel = this;
 
-    if (!MediumClone.router._requireSignedIn(function () {
-      thisModel.fetch({
-        success : function() {
-          if (thisModel.get('following')) {
-            type = 'DELETE';
-          } else {
-            type = 'POST';
-          }
-
-          $.ajax({
-            url : 'api/follow',
-            type : type,
-            data : {
-              follow: {
-                followable_id : thisModel.id,
-                followable_type : thisModel.followableType,
-              }
-            },
-            success : completionCallback,
-          });
-        },
+    if (!MediumClone.router._requireSignedIn(function() {
+      thisModel.toggleFollow(function () {
+        completionCallback();
+        MediumClone.router.showUser(thisModel.id);
+        Backbone.history.navigate('users/' + thisModel.id);
       });
-      MediumClone.router.profile();
-      Backbone.history.navigate('');
     })) { return; }
+
+    thisModel.fetch({
+      success : function() {
+        if (thisModel.get('following')) {
+          type = 'DELETE';
+        } else {
+          type = 'POST';
+        }
+
+        $.ajax({
+          url : 'api/follow',
+          type : type,
+          data : {
+            follow: {
+              followable_id : thisModel.id,
+              followable_type : thisModel.followableType,
+            }
+          },
+          success : completionCallback,
+        });
+      },
+    });
   },
-  
+
 }
